requirements.txt
aiogram==2.23.1
chromedriver-autoinstaller==0.4.0
selenium==4.7.2
requests==2.28.1
python-dotenv==0.21.0
